+++ inst/extdata/sql-queries/v2-od-distritos-clean-csv-view-en.sql
@@ -110,6 +110,5 @@ CREATE OR REPLACE VIEW od_csv_clean AS SELECT
     viajes_km AS trips_total_length_km,
     CAST(year AS INTEGER) AS year,
     CAST(month AS INTEGER) AS month,
-    CAST(day AS INTEGER) AS day,
-    periodo AS time_slot
+    CAST(day AS INTEGER) AS day
     FROM od_csv_raw;